myapp/web/requirements.txt
flask
flask-httpauth
mysql-connector-python
